refactor: data fetch 로직을 util 함수로 분리
